requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
requests>=2.31.0
